try make tests pass on 32-bits machines

+++ stdlib/LinearAlgebra/test/bunchkaufman.jl
@@ -90,7 +90,7 @@ bimg  = randn(n,2)/2
         end
 
         @testset "$eltyb argument B" for eltyb in (Float32, Float64, ComplexF32, ComplexF64, Int)
-            b = eltyb == Int ? rand(1:5, n, 2) : convert(Matrix{eltyb}, eltyb <: Complex ? complex.(breal, bimg) : breal)
+            b = eltyb == Int ? rand(Int64(1):Int64(5), n, 2) : convert(Matrix{eltyb}, eltyb <: Complex ? complex.(breal, bimg) : breal)
             for b in (b, view(b, 1:n, 1:2))
                 εb = eps(abs(float(one(eltyb))))
                 ε = max(εa,εb)
